% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx_modcomp_new.R
\name{modcomp}
\alias{modcomp}
\alias{pb_modcomp}
\alias{kr_modcomp}
\alias{sat_modcomp}
\alias{x2_modcomp}
\title{Compare two models}
\usage{
pb_modcomp(largeModel, smallModel, control = list())

kr_modcomp(largeModel, smallModel, control = list())

sat_modcomp(largeModel, smallModel, control = list())

x2_modcomp(largeModel, smallModel, control = list())
}
\arguments{
\item{largeModel, smallModel}{Two models}

\item{control}{A list}
}
\description{
Compare two models
}
