% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAT_modcomp.R
\name{devfun_vp}
\alias{devfun_vp}
\title{Compute deviance of a linear mixed model as a function of variance
parameters}
\usage{
devfun_vp(varpar, devfun, reml)
}
\arguments{
\item{varpar}{variance parameters; \code{varpar = c(theta, sigma)}.}

\item{devfun}{deviance function as a function of theta only.}

\item{reml}{if \code{TRUE} the REML deviance is computed;
if \code{FALSE}, the ML deviance is computed.}
}
\value{
the REML or ML deviance.
}
\description{
This function is used for extracting the asymptotic
variance-covariance matrix of the variance parameters.
}
\author{
Rune Haubo B. Christensen. Adapted to pbkrtest by Søren Højsgaard.
}
\keyword{internal}
