% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbkr_pb_modcomp.R
\name{summarize_pb}
\alias{summarize_pb}
\title{Summarize parametric bootstrap results for LRT}
\usage{
summarize_pb(LRTstat, ref, conf.level = 0.95)
}
\arguments{
\item{LRTstat}{Numeric vector: c(tobs, df) = observed test statistic and df.}

\item{ref}{Numeric vector of simulated bootstrap test statistics.}

\item{conf.level}{Confidence level for bootstrap p CI (default 0.95).}
}
\value{
List with results table, moments, samples info, CI, SE, etc.
}
\description{
Calculates asymptotic p-value, bootstrap p-value with CI, Bartlett corrections
(mean and median-based) and F-approximations (mean and median-based).
}
