% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbkr_pb_modcomp.R
\name{summarize_pb}
\alias{summarize_pb}
\title{Summarize parametric bootstrap results for LRT}
\usage{
summarize_pb(LRTstat, ref, conf.level = 0.95)
}
\arguments{
\item{LRTstat}{Numeric vector: c(tobs, df) = observed test statistic and df.}

\item{ref}{Numeric vector of simulated bootstrap test statistics.}

\item{conf.level}{Confidence level for bootstrap p CI (default 0.95).}
}
\value{
List with results table, moments, samples info, CI, SE, etc.
}
\description{
Calculates asymptotic p-value, bootstrap p-value with CI, Bartlett corrections
(mean and median-based) and F-approximations (mean and median-based).
}
\examples{

sleepstudy$Days2 <- sleepstudy$Days^2

lm_fit1 <- lm(Reaction ~ Days, data = sleepstudy)
lm_fit0 <- update(lm_fit1, . ~ . - Days)
set.seed(42)
lrt_lm <- getLRT(lm_fit1, lm_fit0)
lr_sim_lm <- pb_refdist(lm_fit1, lm_fit0, nsim = 500, engine = "future", nworkers =2)
summarize_pb(lrt_lm, lr_sim_lm)

gls_fit1 <- gls(Reaction ~ Days, data = sleepstudy, method="ML")
gls_fit0 <- update(gls_fit1, . ~ . - Days)
set.seed(42)
lrt_gls <- getLRT(gls_fit1, gls_fit0)
lr_sim_gls <- pb_refdist(gls_fit1, gls_fit0, nsim = 500, engine = "future", nworkers = 2)
summarize_pb(lrt_gls, lr_sim_gls)

lme_fit1 <- lme(Reaction ~ Days, random = ~ 1 | Subject, data = sleepstudy, method="ML")
lme_fit0 <- update(lme_fit1, . ~ . - Days)
set.seed(42)
lrt_lme <- getLRT(lme_fit1, lme_fit0)
lr_sim_lme <- pb_refdist(lme_fit1, lme_fit0, nsim = 500, engine = "future", nworkers = 2)
summarize_pb(lrt_lme, lr_sim_lme)

lmer_fit1 <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy, REML=FALSE)
lmer_fit0 <- update(lmer_fit1, . ~ . - Days)
set.seed(42)
lrt_lmer <- getLRT(lmer_fit1, lmer_fit0)
lr_sim_lmer <- pb_refdist(lmer_fit1, lmer_fit0, nsim = 500, engine = "future", nworkers = 2)
summarize_pb(lrt_lmer, lr_sim_lmer)

sleepstudy$Days2 <- sleepstudy$Days^2

lm_fit1 <- lm(Reaction ~ Days + Days2, data = sleepstudy)
lm_fit0 <- update(lm_fit1, . ~ . - Days2)
set.seed(42)
lrt_lm <- getLRT(lm_fit1, lm_fit0)
lr_sim_lm <- pb_refdist(lm_fit1, lm_fit0, nsim = 50, engine = "future", nworkers =2)
summarize_pb(lrt_lm, lr_sim_lm)

gls_fit1 <- gls(Reaction ~ Days + Days2, data = sleepstudy, method="ML")
gls_fit0 <- update(gls_fit1, . ~ . - Days2)
set.seed(42)
lrt_gls <- getLRT(gls_fit1, gls_fit0)
lr_sim_gls <- pb_refdist(gls_fit1, gls_fit0, nsim = 50, engine = "future", nworkers = 2)
summarize_pb(lrt_gls, lr_sim_gls)

lme_fit1 <- lme(Reaction ~ Days + Days2, random = ~ 1 | Subject, data = sleepstudy, method="ML")
lme_fit0 <- update(lme_fit1, . ~ . - Days2)
set.seed(42)
lrt_lme <- getLRT(lme_fit1, lme_fit0)
lr_sim_lme <- pb_refdist(lme_fit1, lme_fit0, nsim = 50, engine = "future", nworkers = 2)
summarize_pb(lrt_lme, lr_sim_lme)

lmer_fit1 <- lmer(Reaction ~ Days + Days2 + (1 | Subject), data = sleepstudy, REML=FALSE)
lmer_fit0 <- update(lmer_fit1, . ~ . - Days2)
set.seed(42)
lrt_lmer <- getLRT(lmer_fit1, lmer_fit0)
lr_sim_lmer <- pb_refdist(lmer_fit1, lmer_fit0, nsim = 50, engine = "future", nworkers = 2)
summarize_pb(lrt_lmer, lr_sim_lmer) |> summary()

lr_sim_lmer <- pb_refdist_sequential(lmer_fit1, lmer_fit0, nsim = 50,
    engine = "future", nworkers = 2)
summarize_pb(lrt_lmer, lr_sim_lmer)
}
